LayoutTests/webgpu/webgpu-enabled.html

<!DOCTYPE html><!-- webkit-test-runner [ experimental:WebGPUEnabled=true ] -->
<html>
<meta charset=utf-8>
<title>Test if WebGPU is enabled.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
promise_test(async () => {
    const canvas = document.createElement("canvas");
    const context = canvas.getContext("gpu");
    assert_true(context instanceof GPUCanvasContext, "Created GPUCanvasContext.");

    assert_not_equals(window.gpu, undefined, "window.gpu is defined.");

    const adapter = await window.gpu.requestAdapter();
    assert_true(adapter instanceof WebGPUAdapter, "Created default GPUAdapter.");
    
    const device = adapter.createDevice();
    assert_true(device instanceof WebGPUDevice, "Created GPUDevice.");
}, "Web GPU is enabled.");
</script>
</html>
